Subida completa de los archivos del proyecto (terminado)

+++ 8QueensProyect/Cromosoma.cs
@@ -10,7 +10,7 @@ namespace _8QueensProyect
     {
         public List<int> Genes{ get; set; }
         public int Colisiones { get; set; }
-        public int Fitness { get; set; }
+        public double Fitness { get; set; }
         public int Generacion { get; set; }
         public int Numero { get; set; }
